% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Human_Allele_Frequencies}
\alias{Human_Allele_Frequencies}
\title{Allele frequencies from 19,500 single nucleotide polymorphisms}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 6 rows and 19500 columns.
}
\usage{
Human_Allele_Frequencies
}
\description{
The dataset “Human_Allele_Frequencies” is a 6 × 19,500 matrix
that contains allele frequencies from 19,500 single nucleotide
polymorphisms in Chromosomes 1-10 in 6 human populations.
The last population “San” is intended to be used as an outgroup.
The allele frequencies have been compiled from ALFRED database
at Yale University. The analysis from this dataset has been
published in Peng et al 2021.
}
\keyword{datasets}
